requirements.txt
Django==4.2.7
djangorestframework==3.14.0
model-utils==4.2.0